# Optimize token balances in Rest API