--- a/.github/workflows/rest-api.yml
+++ b/.github/workflows/rest-api.yml
@@ -20,7 +20,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        schema: [ v1, v2 ]
+        schema: [ v1 ]
     timeout-minutes: 18
     steps:
       - name: Checkout repository