--- a/hedera-mirror-rest/__tests__/integrationDomainOps.js
+++ b/hedera-mirror-rest/__tests__/integrationDomainOps.js
@@ -433,7 +433,7 @@ const addEntity = async (defaults, entity) => {
 
 const addEthereumTransaction = async (ethereumTransaction) => {
   // any attribute starting with '_' is not a db column
-  ethereumTransaction = _.omitBy(ethereumTransaction, (v, k) => k.startsWith('_'));
+  ethereumTransaction = _.omitBy(ethereumTransaction, (_v, k) => k.startsWith('_'));
   const localDefaults = {
     access_list: null,
     call_data_id: null,