--- a/hedera-mirror-rest/__tests__/specs/accounts/all-params.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/all-params.json
@@ -56,12 +56,10 @@
         "balance": 70,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 17
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 1717
           }
@@ -80,12 +78,10 @@
         "balance": 90,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 19
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 1919
           }