--- a/hedera-mirror-rest/__tests__/specs/accounts/balance-false.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/balance-false.json
@@ -19,12 +19,10 @@
         "balance": 70,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 100001,
             "balance": 7
           },
           {
-            "token_realm": 0,
             "token_num": 100002,
             "balance": 77
           }
@@ -36,7 +34,6 @@
         "balance": 80,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 200001,
             "balance": 8
           }