--- a/hedera-mirror-rest/__tests__/specs/accounts/specific-id.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/specific-id.json
@@ -23,12 +23,10 @@
         "balance": 70,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 100001,
             "balance": 7
           },
           {
-            "token_realm": 0,
             "token_num": 100002,
             "balance": 77
           }
@@ -40,7 +38,6 @@
         "balance": 80,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 200001,
             "balance": 8
           }