--- a/hedera-mirror-rest/__tests__/specs/accounts/{id}/no-balances.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/{id}/no-balances.json
@@ -27,12 +27,10 @@
         "balance": 20,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 2
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 20
           }
@@ -44,12 +42,10 @@
         "balance": 70,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 7
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 77
           }