--- a/hedera-mirror-rest/__tests__/specs/accounts/{id}/token-balance-limit.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/{id}/token-balance-limit.json
@@ -1,0 +1,141 @@
+{
+  "description": "Account api calls for specific account with token balance limit",
+  "setup": {
+    "config": {
+      "response": {
+        "limit": {
+          "tokenBalance": {
+            "single": 2
+          }
+        }
+      }
+    },
+    "accounts": [
+      {
+        "num": 7
+      },
+      {
+        "num": 8,
+        "alias": "KGNABD5L3ZGSRVUCSPDR7TONZSRY3D5OMEBKQMVTD2AC6JL72HMQ",
+        "evm_address": "ac384c53f03855fa1b3616052f8ba32c6c2a2fec",
+        "public_key": "519a008fabde4d28d68293c71fcdcdcca38d8fae6102a832b31e802f257fd1d9"
+      }
+    ],
+    "balances": [
+      {
+        "timestamp": 2345,
+        "id": 7,
+        "balance": 70,
+        "tokens": [
+          {
+            "token_num": 99997,
+            "balance": 3
+          },
+          {
+            "token_num": 99998,
+            "balance": 7
+          },
+          {
+            "token_num": 99999,
+            "balance": 77
+          }
+        ]
+      },
+      {
+        "timestamp": 2345,
+        "id": 8,
+        "balance": 80,
+        "tokens": [
+          {
+            "token_num": 99998,
+            "balance": 8
+          },
+          {
+            "token_num": 99999,
+            "balance": 88
+          }
+        ]
+      }
+    ]
+  },
+  "tests": [
+    {
+      "url": "/api/v1/accounts/8",
+      "responseStatus": 200,
+      "responseJson": {
+        "transactions": [],
+        "balance": {
+          "timestamp": "0.000002345",
+          "balance": 80,
+          "tokens": [
+            {
+              "token_id": "0.0.99998",
+              "balance": 8
+            },
+            {
+              "token_id": "0.0.99999",
+              "balance": 88
+            }
+          ]
+        },
+        "account": "0.0.8",
+        "alias": "KGNABD5L3ZGSRVUCSPDR7TONZSRY3D5OMEBKQMVTD2AC6JL72HMQ",
+        "decline_reward": false,
+        "deleted": false,
+        "ethereum_nonce": null,
+        "evm_address": "0xac384c53f03855fa1b3616052f8ba32c6c2a2fec",
+        "expiry_timestamp": null,
+        "auto_renew_period": null,
+        "key": null,
+        "max_automatic_token_associations": 0,
+        "memo": "entity memo",
+        "receiver_sig_required": false,
+        "staked_account_id": null,
+        "staked_node_id": null,
+        "stake_period_start": null,
+        "links": {
+          "next": null
+        }
+      }
+    },
+    {
+      "url": "/api/v1/accounts/7",
+      "responseStatus": 200,
+      "responseJson": {
+        "transactions": [],
+        "balance": {
+          "timestamp": "0.000002345",
+          "balance": 70,
+          "tokens": [
+            {
+              "token_id": "0.0.99997",
+              "balance": 3
+            },
+            {
+              "token_id": "0.0.99998",
+              "balance": 7
+            }
+          ]
+        },
+        "account": "0.0.7",
+        "alias": null,
+        "decline_reward": false,
+        "deleted": false,
+        "ethereum_nonce": null,
+        "evm_address": null,
+        "expiry_timestamp": null,
+        "auto_renew_period": null,
+        "key": null,
+        "max_automatic_token_associations": 0,
+        "memo": "entity memo",
+        "receiver_sig_required": false,
+        "staked_account_id": null,
+        "staked_node_id": null,
+        "stake_period_start": null,
+        "links": {
+          "next": null
+        }
+      }
+    }
+  ]
+}