--- a/hedera-mirror-rest/__tests__/specs/accounts/{id}/token-balance-limit.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/{id}/token-balance-limit.json
@@ -5,7 +5,7 @@
       "response": {
         "limit": {
           "tokenBalance": {
-            "single": 2
+            "singleAccount": 2
           }
         }
       }