--- a/hedera-mirror-rest/__tests__/specs/accounts/{id}/transactiontype.json
+++ b/hedera-mirror-rest/__tests__/specs/accounts/{id}/transactiontype.json
@@ -27,12 +27,10 @@
         "balance": 70,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 7
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 77
           }
@@ -44,12 +42,10 @@
         "balance": 80,
         "tokens": [
           {
-            "token_realm": 0,
             "token_num": 99998,
             "balance": 8
           },
           {
-            "token_realm": 0,
             "token_num": 99999,
             "balance": 88
           }