--- a/hedera-mirror-rest/__tests__/specs/balances/limit.json
+++ b/hedera-mirror-rest/__tests__/specs/balances/limit.json
@@ -7,7 +7,7 @@
           "default": 2,
           "max": 4,
           "tokenBalance": {
-            "multiple": 2
+            "multipleAccounts": 2
           }
         }
       }