**Description**:
<!--
One or two line summary of what this PR does and why it is needed, followed by a list
of changes in imperative, present tense for use in the commit message or changelog. Example:

This PR modifies ... in order to support ...
* Add config property
* Change column name
* Remove ...
-->

**Related issue(s)**:

Fixes #4131 

**Notes for reviewer**:
<!-- Provide logs, performance numbers or screenshots of the new functionality -->

Had to disable v2 schema rest api workflow since citus doesn't support such query: `correlated subqueries are not supported when the FROM clause contains a CTE or subquery`

k6 report:

- `main` branch

| URL | VUS | Pass% | RPS | Pass RPS | Avg. Req Duration | Comment |
|----------|-----|-------|-----|----------|-------------------|---------|
| /accounts/{id}/allowances/crypto | 500 | 100.00 | 3116.68/s | 3116.68/s | 159.76ms | |
| /accounts/{id}/allowances/crypto | 500 | 100.00 | 3072.10/s | 3072.10/s | 162.07ms | |
| /accounts | 500 | 100.00 | 378.55/s | 378.55/s | 858.24ms | |
| /accounts?balance=false | 500 | 100.00 | 1483.73/s | 1483.73/s | 246.78ms | |
| /accounts?balance=false&account.publickey={publicKey} | 500 | 100.00 | 2484.76/s | 2484.76/s | 200.39ms | |
| /accounts?account.balance=gt:0 | 500 | 100.00 | 1341.87/s | 1341.87/s | 256.70ms | |
| /accounts?account.balance=ne:{balance}&order=desc | 500 | 100.00 | 1343.04/s | 1343.04/s | 259.54ms | |
| /accounts/{accountId} | 500 | 100.00 | 2290.16/s | 2290.16/s | 217.42ms | |
| /accounts?account.id=ne:{accountId}&order=desc | 500 | 100.00 | 1464.15/s | 1464.15/s | 243.06ms | |
| /balances | 500 | 100.00 | 1582.48/s | 1582.48/s | 267.06ms | |

- PR branch

| URL | VUS | Pass% | RPS | Pass RPS | Avg. Req Duration | Comment |
|----------|-----|-------|-----|----------|-------------------|---------|
| /accounts/{id}/allowances/crypto | 500 | 100.00 | 3133.02/s | 3133.02/s | 158.94ms | |
| /accounts/{id}/allowances/crypto | 500 | 100.00 | 3091.40/s | 3091.40/s | 161.02ms | |
| /accounts | 500 | 100.00 | 379.27/s | 379.27/s | 866.80ms | |
| /accounts?balance=false | 500 | 100.00 | 1476.24/s | 1476.24/s | 245.94ms | |
| /accounts?balance=false&account.publickey={publicKey} | 500 | 100.00 | 2516.38/s | 2516.38/s | 197.85ms | |
| /accounts?account.balance=gt:0 | 500 | 100.00 | 1342.99/s | 1342.99/s | 251.94ms | |
| /accounts?account.balance=ne:{balance}&order=desc | 500 | 100.00 | 1336.33/s | 1336.33/s | 260.16ms | |
| /accounts/{accountId} | 500 | 100.00 | 2387.00/s | 2387.00/s | 208.69ms | |
| /accounts?account.id=ne:{accountId}&order=desc | 500 | 100.00 | 1460.37/s | 1460.37/s | 247.47ms | |
| /balances | 500 | 100.00 | 1593.69/s | 1593.69/s | 263.34ms | |

There is no noticeable performance difference. 

**Checklist**

- [x] Documented (Code comments, README, etc.)
- [x] Tested (unit, integration, etc.)